% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mech_names.r
\name{clean_mech_names}
\alias{clean_mech_names}
\title{clean the mechanism categories replacing space and other symbols
 with _.}
\usage{
clean_mech_names(...)
}
\arguments{
\item{x}{input variable}
}
\value{
return cleaner mechanism categories
}
\description{
clean the mechanism categories replacing space and other symbols
 with _.
}
\examples{
library(tidyverse)
clean_mech_names(c("Natural/Environmental, Other"))


}
