% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intent_mech.r
\name{create_intent_mech}
\alias{create_intent_mech}
\title{Add intent and mechanism combined fields for injury ICD-10-CM.}
\usage{
create_intent_mech(data, inj_col, reference = c("both", "intent",
  "mechanism"))
}
\arguments{
\item{data}{input data}

\item{inj_col}{ecode and diagnosis column indices}

\item{reference}{injury matrix reference with three possible choices:
"both"(the default) adding intent and mechanism combined variables
"intent" adding intent only
"mechanism" adding mechanism only}
}
\value{
return the input with additional variables (93 intent_mechanism combinations, five intent variables, or 32 mechanism variable depending on the reference choice
}
\description{
, reference
}
\examples{
library(tidyverse)
dat <- data.frame(d1 = c("T63023", "X92821", "X99100", "T360x"),
d2 = c("T65823", "Y030x0", "T17200", "V0100x" ))

dat \%>\% intent_mech(inj_col = c(1,2), reference = "both")
dat \%>\% intent_mech(inj_col = c(1,2), reference = "intent")
dat \%>\% intent_mech(inj_col = c(1,2), reference = "mechanism")

}
