% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm_codes_2019_regex}
\alias{icd10cm_codes_2019_regex}
\title{icd-10-cm list and descriptions.}
\format{Data frame}
\source{
\url{https://www.cdc.gov/nchs/icd/icd10cm.htm}
}
\usage{
icd10cm_codes_2019_regex
}
\description{
Dataset of 143,865 rows and 3 variables.
}
\examples{
tail(icd10cm_codes_2019_regex)

}
\keyword{datasets}
