% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm_data150}
\alias{icd10cm_data150}
\title{Dataset with icd-10-cm codes.}
\format{Data frame}
\source{
created to use in examples.
}
\usage{
icd10cm_data150
}
\description{
Dataset of 150 rows and 2 variables.
}
\examples{
library(dplyr)
icd10cm_data150 \%>\% add_inj_intent_mech_icd10cm(icd10cm_main = principal_diag)
data_mech <- icd10cm_data150 \%>\%
add_inj_intent_mech_icd10cm(icd10cm_main = principal_diag, reference = "mechanism")
data_mech \%>\% icd10cm_data150 \%>\%
add_inj_intent_mech_icd10cm(icd10cm_main = principal_diag, reference = "mechanism") \%>\%
drop_na() \%>\%
count(mechanism) \%>\% arrange(-n) \%>\%
top_n(10)
}
\keyword{datasets}
