% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm_injury_matrix}
\alias{icd10cm_injury_matrix}
\title{The complete icd10cm injury matrix.}
\format{Data frame}
\source{
\url{https://www.cdc.gov/injury/wisqars/dataandstats.html}
}
\usage{
icd10cm_injury_matrix
}
\description{
Dataset of 3,522 rows and 4 variables.
formatted from the original
}
\examples{
library(dplyr)
sample_n(icd10cm_injury_matrix, 10)

}
\keyword{datasets}
