% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm_intent_mech_regex}
\alias{icd10cm_intent_mech_regex}
\title{The icd10cm injury matrix by intent and mechanism.}
\format{Data frame}
\source{
modified from the injury matrix "icd10cm_injury_matrix"
}
\usage{
icd10cm_intent_mech_regex
}
\description{
Dataset of 93 rows and 4 variables.
grouped by intent and mechanism from original with icd10cm code in regular expressions
}
\examples{
head(icd10cm_intent_mech_regex, 10)

}
\keyword{datasets}
