% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm_intent_regex}
\alias{icd10cm_intent_regex}
\title{icd10cm injury matrix by intent only.}
\format{Data frame}
\source{
Grouped from icd10cm_inj_regex
}
\usage{
icd10cm_intent_regex
}
\description{
Dataset of 6 rows and 2 variables.
}
\examples{
icd10cm_intent_regex

}
\keyword{datasets}
