% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm_mech_regex}
\alias{icd10cm_mech_regex}
\title{icd10cm injury matrix by mechanism only.}
\format{Data frame}
\source{
Grouped from icd10cm_inj_regex.
}
\usage{
icd10cm_mech_regex
}
\description{
Dataset of 34 rows and 2 variables.
}
\examples{
library(dplyr)
sample_n(icd10cm_mech_regex, 10)

}
\keyword{datasets}
