% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regex.R
\name{make_regex}
\alias{make_regex}
\title{transform the ICD-10-CM codes into an operational regular expression.}
\usage{
make_regex(...)
}
\arguments{
\item{x}{input variable}
}
\value{
collapse the grouped icd10cm as regular expressions with all the "x" not at the beginning of the icd10cm code replaced by '.' to match any character.
}
\description{
transform the ICD-10-CM codes into an operational regular expression.
}
\examples{
to be added
library(tidyverse)
make_regex(c("X45x2", "Y65xx"))

}
