% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm_data150}
\alias{icd10cm_data150}
\title{Dataset with icd-10-cm codes.}
\format{Data frame}
\source{
created to use in examples.
}
\usage{
icd10cm_data150
}
\description{
Dataset of 150 rows and 3 variables.
}
\examples{
icd10cm_data150

}
\keyword{datasets}
